packages/api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = ""
readme = "README.md"
authors = [
    { name = "nayetdet", email = "jpmoreiracampos21@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.16.4",
    "chromadb>=1.0.20",
    "fastapi>=0.116.1",
    "llama-index>=0.13.2",
    "llama-index-embeddings-ollama>=0.8.1",
    "llama-index-llms-ollama>=0.7.1",
    "llama-index-vector-stores-chroma>=0.5.0",
    "minio>=7.2.16",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.43",
    "uvicorn>=0.35.0",
]

[project.scripts]
api = "api:main"

[build-system]
requires = ["uv_build>=0.8.4,<0.9.0"]
build-backend = "uv_build"
